+++ bugsnag/utils.py
@@ -277,41 +277,27 @@ def merge_dicts(lhs, rhs):
             lhs[key] = value
 
 
-class ThreadLocals(object):
+class ThreadContextVar:
+    """
+    A wrapper around thread-local variables to mimic the API of contextvars
+    """
     LOCALS = None
 
-    @staticmethod
-    def get_instance():
-        if not ThreadLocals.LOCALS:
-            ThreadLocals.LOCALS = threadlocal()
-        return ThreadLocals()
+    @classmethod
+    def local_context(cls):
+        if not ThreadContextVar.LOCALS:
+            ThreadContextVar.LOCALS = threadlocal()
+        return ThreadContextVar.LOCALS
 
-    def get_item(self, key, default=None):
-        return getattr(ThreadLocals.LOCALS, key, default)
-
-    def set_item(self, key, value):
-        return setattr(ThreadLocals.LOCALS, key, value)
-
-    def has_item(self, key):
-        return hasattr(ThreadLocals.LOCALS, key)
-
-    def del_item(self, key):
-        return delattr(ThreadLocals.LOCALS, key)
-
-
-class ThreadContextVar(object):
-    """
-    A wrapper around ThreadLocals to mimic the API of contextvars
-    """
     def __init__(self, name, default=None):
         self.name = name
-        ThreadLocals.get_instance().set_item(name, default)
+        setattr(ThreadContextVar.local_context(), name, default)
 
     def get(self):
-        local = ThreadLocals.get_instance()
-        if local.has_item(self.name):
-            return local.get_item(self.name)
+        local = ThreadContextVar.local_context()
+        if hasattr(local, self.name):
+            return getattr(local, self.name)
         raise LookupError("No value for '{}'".format(self.name))
 
     def set(self, new_value):
-        ThreadLocals.get_instance().set_item(self.name, new_value)
+        setattr(ThreadContextVar.local_context(), self.name, new_value)